+++ net-spider/ChangeLog.md
@@ -17,6 +17,13 @@
 
 * Add `FromJSON` and `ToJSON` instances to `LinkState`, `FoundLink`
   and `FoundNode`.
+* Add `Ord` instance to `FoundNode`.
+
+### Snapshot module
+
+* Add `FromJSON` and `ToJSON` instances to `SnapshotLink` and
+
+  `SnapshotNode`.
 
 ## 0.4.0.1  -- 2019-12-29